requirements/base.txt
Django==3.1.3
django-environ==0.4.5
psycopg2-binary==2.8.6
django-braces==1.14.0
pypugjs==5.9.5
django-crispy-forms==1.9.2
django-datatables-view==1.19.1
celery==4.4.7
django-celery-results==1.2.1
django-celery-beat==2.1.0
redis==3.5.3
flower==0.9.5
channels==2.4.0
channels-redis==2.4.2
social-auth-app-django==4.0.0
django-cleanup==5.1.0
djangorestframework==3.12.2
Markdown=3.3.3
django-filter==2.4.0
